Restructured sync logic and added informative UI on sync status as well as option to retrigger sync

+++ mobile-app/lib/backend/storage/storage_repository.dart
@@ -2,12 +2,48 @@ import 'dart:io';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:amplify_storage_s3/amplify_storage_s3.dart';
 import 'package:connectivity_plus/connectivity_plus.dart';
+import 'package:flutter/services.dart';
+import 'package:mobile_app/backend/Blocs/sync/sync_bloc.dart';
+import 'package:mobile_app/backend/Blocs/sync/sync_events.dart';
 import 'package:mobile_app/backend/repositories/LocalDataRepository.dart';
 import 'package:mobile_app/utils/connectivity.dart';
+import 'package:path_provider/path_provider.dart';
 
 import 'dataStorePaths.dart';
 
 class StorageRepository {
+  static SyncBloc? syncBloc;
+
+  static Future<bool> hasConnectivity() async {
+    if (_last_connectivity_check != null) {
+      if (_last_connectivity_check!
+              .difference(DateTime.now())
+              .inMilliseconds
+              .abs() <
+          500) {
+        if (_had_connectivity != null) {
+          return Future.value(_had_connectivity);
+        }
+      }
+    }
+
+    try {
+      final result = await InternetAddress.lookup('google.com');
+      if (result.isNotEmpty && result[0].rawAddress.isNotEmpty) {
+        _had_connectivity = true;
+      } else {
+        _had_connectivity = false;
+      }
+    } on SocketException catch (_) {
+      _had_connectivity = false;
+    }
+    _last_connectivity_check = DateTime.now();
+    return _had_connectivity!;
+  }
+
+  static bool? _had_connectivity;
+  static DateTime? _last_connectivity_check;
+
   static Future<void> downloadFile(File toDownload, String path,
       {bool checkConnection = true, DateTime? lastModifiedOnline}) async {
     try {
@@ -16,34 +52,43 @@ class StorageRepository {
         InternetConnectionType internetConnectionType =
             await StorageRepository.currentInternetConnectionType();
         if (LocalDataRepository.instance.wifiOnly &&
-        internetConnectionType != InternetConnectionType.WIFI) {
+            internetConnectionType != InternetConnectionType.WIFI) {
           return;
         }
       }
 
+      syncBloc?.add(StartLoadingFileEvent());
+
+      //check internet connection
+      if (!await hasConnectivity()) {
+        syncBloc?.add(CancelSyncEvent());
+        return;
+      }
+
       await Amplify.Storage.downloadFile(key: path, local: toDownload);
       //set local last changed to onlyine last changed
+      syncBloc?.add(LoadedFileEvent());
 
-      if(lastModifiedOnline!=null) {
+      if (lastModifiedOnline != null) {
         await toDownload.setLastModified(lastModifiedOnline);
-      }
-      else {
+      } else {
         try {
           ListResult listResult = await Amplify.Storage.list(path: path);
-          if(listResult.items.isNotEmpty) {
+          if (listResult.items.isNotEmpty) {
             lastModifiedOnline = listResult.items.first.lastModified;
-            if(lastModifiedOnline!=null) {
+            if (lastModifiedOnline != null) {
               await toDownload.setLastModified(lastModifiedOnline);
             }
           }
-        } catch(e) {
+        } catch (e) {
           print('Error in catching last modified online');
         }
       }
 
-
       print("file for $path successfully downloaded");
     } catch (e) {
+      syncBloc?.add(LoadedFileEvent());
+      //todo: handle with sync bloc
       print("not found $path");
       print(e.toString());
     }
@@ -61,13 +106,27 @@ class StorageRepository {
           return dataStorePath;
         }
       }
+
+      //add to sync bloc
+      syncBloc?.add(StartLoadingFileEvent());
+
+      if (!await hasConnectivity()) {
+        syncBloc?.add(CancelSyncEvent());
+        return dataStorePath;
+      }
+
       final result = await Amplify.Storage.uploadFile(
         local: file,
         key: dataStorePath,
       );
 
+      //remove from sync bloc
+      syncBloc?.add(LoadedFileEvent());
+
       return result.key;
     } catch (e) {
+      syncBloc?.add(LoadedFileEvent());
+      //remove from sync bloc
       return dataStorePath;
     }
   }
@@ -85,12 +144,26 @@ class StorageRepository {
     try {
       // RemoveOptions options =
       // RemoveOptions(accessLevel: StorageAccessLevel.guest);
+
+      //send to sync bloc
+      syncBloc?.add(StartLoadingFileEvent());
+
+      if (!await hasConnectivity()) {
+        syncBloc?.add(CancelSyncEvent());
+        return path;
+      }
+
       final result = await Amplify.Storage.remove(
         key: path,
         // options: options
       );
+
+      syncBloc?.add(LoadedFileEvent());
+
       return result.key;
     } catch (e) {
+      //remove from sync bloc
+      syncBloc?.add(LoadedFileEvent());
       rethrow;
     }
   }
@@ -115,6 +188,32 @@ class StorageRepository {
     } on SocketException catch (_) {
       //no connection
       return InternetConnectionType.OFFLINE;
+    }
+  }
+
+  static Future<bool> dbObjectSave(
+      Map<String, dynamic> values, String id, String type) async {
+    String path = dataStorePath(DataStorePaths.failedDBObject, [type, id]);
+
+    try {
+      //upload json files from values
+
+      //create a file in cache from values without a constant path -> will be deleted afterwards
+      Directory appDocDir = await getApplicationDocumentsDirectory();
+      Directory(appDocDir.path).createSync(recursive: true);
+      File localCacheFile = File('${appDocDir.path}/$path');
+      localCacheFile.parent.createSync(recursive: true);
+      await localCacheFile.writeAsString(values.toString(), flush: true);
+      await Amplify.Storage.uploadFile(
+        local: localCacheFile,
+        key: path,
+      );
+      await localCacheFile.delete();
+      return true;
+    } catch (e) {
+      print('error in db save operation');
+      print(e.toString());
+      return false;
     }
   }
 }
